fix: make semantic planning target neutral
