+++ src/render/conversion-renderer.ts
@@ -3,6 +3,7 @@ import type { PlannedConversion } from '../adapter/conversion-adapter';
 import type { BreakpointMigrationConfig } from '../config/breakpoint-migration-config';
 import type { SemanticConversionContext } from '../semantic/conversion-context';
 import type { ResolvedSemanticPlan } from '../semantic/semantic-plan';
+import type { SourcePropertyEvidence } from '../semantic/source-property-evidence';
 
 export interface RenderedConversion {
   readonly status: 'converted';
@@ -13,6 +14,7 @@ export interface RenderedConversion {
 export interface ConversionRenderer {
   readonly target: 'tailwind' | 'css';
   readonly breakpointConfig?: BreakpointMigrationConfig;
+  readonly sourcePropertyEvidence?: SourcePropertyEvidence;
   eligibility(input: LocatedFlexLayoutInput): PlannedConversion | undefined;
   render(plan: ResolvedSemanticPlan, context: SemanticConversionContext): PlannedConversion;
   resolveConflicts(
